kube-aws: Drain nodes before shutting them down to give running pods time to gracefully stop.
This change basically achieves it by running `docker run IMAGE kubectl drain THE_NODE --force` on the to-be-shut-down node before the kubelet gets SIGTERM in a CoreOS' shutdown process.

Without this change, kubelets getting SIGTERM without a prior `drain` results in unfunctional(actually not schedulable) pods to have statuses `Ready`.

With this change, when an ASG's desired cap decreased, a node's status changes over time as follows:

On desired cap change:
  STATUS=Ready
On shut-down started:
  STATUS=Ready,SchedulingDisabled (<- Pods are stopped and status is changed by `kubectl drain`)
On shut-down finished:
  Status=NotReady,SchedulingDisabled (<- It's `NotReady` but it won't result in a down-time because we already stopped both further scheduling and pods)
After a minute:
  The node disappears from the output of `kubectl get nodes`

Note that:

* This applies to manual shutdowns(via running `sudo systemctl shutdown` for example) and automated shutdowns(triggered by AWS AutoScaling when nodes get rotated out of a group)
* We currently depend on the community docker image `mumoshu/kubectl` because `kubectl` included in the official `coreos/hyperkube` image doesn't work due to the issue https://github.com/kubernetes/kubernetes/issues/24088 in Kubernetes. Once the issue is fixed and the CoreOS team published the new hyperkube image with the updated Kubernetes, we can remove that dependency.
* The author considers this an experimental feature. So you shouldn't expect configuration API regarding this stable. It may change in the future.

Re-format code introduced in the previous with gofmt to conform the rules and make the build pases